Add balance to user page and some refactoring in the codebase

+++ app/src/main/java/ch/epfl/favo/user/User.java
@@ -57,7 +57,7 @@ public class User implements Document {
   private int completedFavors;
   private int likes;
   private int dislikes;
-  private double balance;
+  private int balance;
   private String profilePictureUrl;
   private double notificationRadius;
   private boolean chatNotifications;
@@ -71,7 +71,7 @@ public class User implements Document {
     this.likes = 0;
     this.dislikes = 0;
     this.completedFavors = 0;
-    this.balance = 10.0;
+    this.balance = 10;
     this.notificationRadius = 10.0;
     this.updateNotifications = true;
     this.chatNotifications = true;
@@ -92,7 +92,7 @@ public class User implements Document {
     this.completedFavors = (int) map.get(COMPLETED_FAVORS);
     this.likes = (int) map.get(LIKES);
     this.dislikes = (int) map.get(DISLIKES);
-    this.balance = (double) map.get(BALANCE);
+    this.balance = (int) map.get(BALANCE);
     this.profilePictureUrl = (String) map.get(PROFILE_PICTURE_URL);
     this.notificationRadius = (double) map.get(NOTIFICATION_RADIUS);
     this.chatNotifications = (boolean) map.get(CHAT_NOTIFICATIONS);
@@ -120,7 +120,7 @@ public class User implements Document {
     this.likes = 0;
     this.dislikes = 0;
     this.completedFavors = 0;
-    this.balance = 10.0;
+    this.balance = 10;
     this.notificationRadius = 10.0;
     this.updateNotifications = true;
     this.chatNotifications = true;
@@ -197,7 +197,7 @@ public class User implements Document {
     return notificationId;
   }
 
-  public double getBalance() {
+  public int getBalance() {
     return balance;
   }
 
@@ -241,7 +241,7 @@ public class User implements Document {
     this.notificationId = notificationId;
   }
 
-  void setBalance(Double balance) {
+  void setBalance(int balance) {
     this.balance = balance;
   }
