+++ app/src/main/java/ch/epfl/favo/user/UserUtil.java
@@ -116,7 +116,7 @@ public class UserUtil implements IUserUtil {
   }
 
   @Override
-  public CompletableFuture<Void> updateCoinBalance(String userId, double reward) {
+  public CompletableFuture<Void> updateCoinBalance(String userId, int reward) {
     return findUser(userId)
         .thenCompose(
             (user) -> {
